'change helicity False' as suggested by authors
From private conversation with Olivier Mattalear:

Note that if you use version before 2.5.1, I should advise you to always have the following line in the reweight_card
change helicity False
in the default model “change helicity True", they are a bug (that I wrongly fix in 2.4.3 and correctly fixed in 2.5.1).
The problem in 2.3.3 and 2.4.2 is that the helicity are evaluated in the wrong frame which can lead to problem in some cases.

change helicity False
correspond to equation (1) of https://arxiv.org/pdf/1607.00763.pdf
while change helicity True is equation (4)