fetched jobs ids
Dear Andreas,
I noticed that the finished jobs do not removed from condor_q even after condor_transfer_data call.
Because of this we are retrieving output multiple times for the same job and maybe load the server.
In this patch I add a self.fetched_jobs list in order to store ids of finished and fetched jobs and skip them after. I'm also still have a "condor_transfer_data:" stderr (condor_transfer_data return code 1 in this case) but the number of this errors is much much less.